Rename entities to 'Musicas' and 'Artistas'
