+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/controller/MusicController.java
@@ -1,17 +1,26 @@
 package com.ciandt.summit.bootcamp2022.controller;
 
+import com.ciandt.summit.bootcamp2022.service.MusicService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/v1/music")
+@RequestMapping(value = "/api/v1/music")
 public class MusicController {
+
+    @Autowired
+    private MusicService service;
 
     @GetMapping
     public ResponseEntity<String> get() {
         return ResponseEntity.ok("67f5976c-eb1e-404e-8220-2c2a8a23be47");
     }
 
+    @GetMapping("/all")
+    public ResponseEntity getAll() {
+        return ResponseEntity.ok(service.getAll());
+    }
 }
