--- MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/model/Artist.java
@@ -1,38 +1,0 @@
-package com.ciandt.summit.bootcamp2022.model;
-
-import lombok.*;
-import org.hibernate.Hibernate;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import java.util.Objects;
-
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "Artistas")
-public class Artist {
-    @Id
-    @Column(name = "Id")
-    private String id;
-    @Column(name = "Nome")
-    private String name;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        Artist artist = (Artist) o;
-        return id != null && Objects.equals(id, artist.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return getClass().hashCode();
-    }
-}
