--- MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/model/Music.java
@@ -1,39 +1,0 @@
-package com.ciandt.summit.bootcamp2022.model;
-
-import lombok.*;
-import org.hibernate.Hibernate;
-
-import javax.persistence.*;
-import java.util.Objects;
-
-@Getter
-@Setter
-@ToString
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "Musicas")
-public class Music {
-    @Id
-    @Column(name = "Id")
-    private String id;
-    @Column(name = "Nome")
-    private String name;
-
-    @ManyToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name = "ArtistaId")
-    private Artist artist;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        Music music = (Music) o;
-        return id != null && Objects.equals(id, music.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return getClass().hashCode();
-    }
-}
