+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/repository/MusicRepository.java
@@ -1,6 +1,6 @@
 package com.ciandt.summit.bootcamp2022.repository;
 
-import com.ciandt.summit.bootcamp2022.model.Music;
+import com.ciandt.summit.bootcamp2022.entity.Musicas;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -9,10 +9,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MusicRepository extends JpaRepository<Music, String> {
+public interface MusicRepository extends JpaRepository<Musicas, String> {
 
-    public List<Music> findByArtistaId(String id);
-    public List<Music> findByNomeContainingIgnoreCase(String nome);
+    public List<Musicas> findByArtistId(String id);
+    public List<Musicas> findByNameContainingIgnoreCase(String name);
     @Query("SELECT m FROM Music m JOIN Artist a on a.id = m.artist WHERE lower(a.name) like lower(concat('%', :name,'%')) or lower(m.name) like lower(concat('%', :name,'%')) ORDER BY a.name, m.name ASC")
-    public List<Music> findByNameArtistOrNameMusic(@Param("name") String name);
+    public List<Musicas> findByNameArtistOrNameMusic(@Param("name") String name);
 }
