+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/service/MusicService.java
@@ -1,0 +1,19 @@
+package com.ciandt.summit.bootcamp2022.service;
+
+import com.ciandt.summit.bootcamp2022.entity.Musicas;
+import com.ciandt.summit.bootcamp2022.repository.MusicRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class MusicService {
+
+    @Autowired
+    private MusicRepository musicRepository;
+
+    public List<Musicas> getAll() {
+        return musicRepository.findAll();
+    }
+}
